Delay initial dump of K4 parameters until ready
We had a race condition where sometime the initial dump of K4 parameters
was received before our widgets had subscribed to them. This caused the
vfo widgets to sometimes not be initialized. We now delay telling the K4
that we're ready for the dump until we're actually ready.

I hope this fixes the issue we were seeing.