Fix Exit button operation on lost connection
There were two problems. First the directive, QUIT!, was missing the
exclamation point.  The second was that after losing connectins, we
correctly didn't send commands to the K4. Incorrectly, we also didn't
act on internal directives like QUIT!.

For good measure, instead of calling the quit() function, we instead
call the root.destroy() if tkinter is running, and we call sys.exit()
if tkinter is not running.