Fix bugs in equalizer code
There were a few. In particular, it would previously unconditionally
set the K4's equalizer settings on startup. Now, we've added an
Enable/Disable button. When disabled (default), no equalizer changes
are sent to eht K4. When enabled, changes are still unconditionally
sent at startup according to the equalizer settings in k4companion.

This is a substantial rewrite of the equalizer code.