Handle a PC response in K3 legacy mode format
Apparently sometimes the K4 returns a PC response in legacy mode format.
I don't think this is supposed to happen because the first command we
send to the K4 is "K41;", which requests that all responses be in K4
mode format. So, this appears to be a minor K4 bug. We'll essentially
ignore PC responses in the older format.