Add support for multiple radios
Added a new option "-r" or "--radio" that takes the name
of a radio. This option is implemented by using a different
settings file for each radio. The name of the settings file
is .../<radioname>-settings.ini. For example, on Linux, the
settings file for the radio K4one would be
~/.configs/k4companion/K4one-settings.ini.