bugfix: Do not re-send a frequency rcvd from the K4
We were echoing back change frequency commands (FA$) received from the
K4. This caused problems. We no longe do so.

Also fix the lower frequency limit. The K4 caught this, so no harm
was caused, but now it's fixed.

Also fix the format of the sent FA$ commands. Again, the K4 accepted
what we sent, but now we conform to the spec.