Keep the CW Tx queue as full as possible
To receive accurate feedback from the K4 on how many characters are in
its CW Tx queue, we are limited to a maximum of 9 characters in its
queue.

We'll try to keep up to 9 characters in the K4's CW tx queue, even if
that means that we'll normally be sending 1 character at a time to the
K4.