Output internal YAML file given the --yaml option
If YAML code is contained within the config_data variable and
no YAML config_file is specified on the command line, then output
the internal YAML code when the --yaml command line argument is given.

If a YAML file is specified on the command line, the --yaml command line
argument will cause the preprocessed YAML code to be output.